Rename a file + fix typo
