+++ pkg/controller/elasticsearch/driver/upgrade_predicates.go
@@ -5,11 +5,8 @@
 package driver
 
 import (
-	"sort"
-
 	"github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/client"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/label"
-	"github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/sset"
 	corev1 "k8s.io/api/core/v1"
 )
 
@@ -40,36 +37,23 @@ func NewPredicateContext(
 	}
 }
 
-// sortCandidates is the default sort function, masters have lower priority as
-// we want to update the data nodes first.
-// If 2 Pods are of the same type then use the reverse ordinal order.
-// TODO: Add some priority to unhealthy (bootlooping) Pods
-func sortCandidates(allPods []corev1.Pod) {
-	sort.Slice(allPods, func(i, j int) bool {
-		pod1 := allPods[i]
-		pod2 := allPods[j]
-		if (label.IsMasterNode(pod1) && label.IsMasterNode(pod2)) ||
-			(!label.IsMasterNode(pod1) && !label.IsMasterNode(pod2)) { // same type, use the reverse name function
-			ssetName1, ord1, err := sset.StatefulSetName(pod1.Name)
-			if err != nil {
-				return false
+func applyPredicates(ctx PredicateContext, candidates []corev1.Pod, maxUnavailableReached bool, allowedDeletions int) (deletedPods []corev1.Pod, err error) {
+	for _, candidate := range candidates {
+		if ok, err := runPredicates(ctx, candidate, deletedPods, maxUnavailableReached); err != nil {
+			return deletedPods, err
+		} else if ok {
+			candidate := candidate
+			// Remove from healthy nodes if it was there
+			delete(ctx.healthyPods, candidate.Name)
+			// Append to the deletedPods list
+			deletedPods = append(deletedPods, candidate)
+			allowedDeletions--
+			if allowedDeletions <= 0 {
+				break
 			}
-			ssetName2, ord2, err := sset.StatefulSetName(pod2.Name)
-			if err != nil {
-				return false
-			}
-			if ssetName1 == ssetName2 {
-				// same name, compare ordinal, higher first
-				return ord1 > ord2
-			}
-			return ssetName1 < ssetName2
 		}
-		if label.IsMasterNode(pod1) && !label.IsMasterNode(pod2) {
-			// pod2 has higher priority since it is a data node
-			return false
-		}
-		return true
-	})
+	}
+	return deletedPods, nil
 }
 
 var predicates = [...]Predicate{
@@ -118,7 +102,7 @@ var predicates = [...]Predicate{
 			maxUnavailableReached bool,
 		) (b bool, e error) {
 			// Green health is retrieved only once from the cluster.
-			// We rely on "shard conflict" predicate to avoid to delete to ES nodes that share some shards.
+			// We rely on "shard conflict" predicate to avoid to delete two ES nodes that share some shards.
 			green, err := context.esState.GreenHealth()
 			if err != nil {
 				return false, err
